+++ transactions/common/schema/__init__.py
@@ -40,6 +40,9 @@ _, TX_SCHEMA_CHAIN_MIGRATION_ELECTION = _load_schema("transaction_chain_migratio
 
 _, TX_SCHEMA_VOTE = _load_schema("transaction_vote", TX_SCHEMA_VERSION)
 
+_, TX_SCHEMA_COMPOSE = _load_schema("transaction_compose", TX_SCHEMA_VERSION)
+
+_, TX_SCHEMA_DECOMPOSE = _load_schema("transaction_decompose", TX_SCHEMA_VERSION)
 
 TX_SCHEMA_PATH_2_0, TX_SCHEMA_COMMON_2_0 = _load_schema("transaction", TX_SCHEMA_VERSION_2_0)
 _, TX_SCHEMA_CREATE_2_0 = _load_schema("transaction_create", TX_SCHEMA_VERSION_2_0)
@@ -88,8 +91,12 @@ def validate_transaction_schema(tx):
             _validate_schema(TX_SCHEMA_COMMON, tx)
             if tx["operation"] == "TRANSFER":
                 _validate_schema(TX_SCHEMA_TRANSFER, tx)
-            else:
+            elif tx["operation"] == "CREATE":
                 _validate_schema(TX_SCHEMA_CREATE, tx)
+            elif tx["operation"] == "COMPOSE":
+                _validate_schema(TX_SCHEMA_COMPOSE, tx)
+            elif tx["operation"] == "DECOMPOSE":
+                _validate_schema(TX_SCHEMA_DECOMPOSE, tx)
         else:
             _validate_schema(TX_SCHEMA_COMMON_2_0, tx)
             if tx["operation"] == "TRANSFER":