+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.6.0] - 2022-01-26
+### Added
+- Transactional types: Compose and Decompose
+- structural tests for Transaction class
+
 ## [0.5.0] - 2022-12-12
 ### Fixed
 - renamed method from _simple_ to _ed25519_ (a more expressive and speaking name)